pyproject.toml
[tool.poetry]
name = "fipe"
version = "0.1.0"
description = ""
authors = ["Patreze Lopes <patreze@lopes.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.10.4"
sqlalchemy = "^2.0.12"
fastapi = "^0.95.1"
celery = "^5.2.7"
python-decouple = "^3.8"
pika = "^1.3.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
requests = "^2.30.0"
pytest = "^7.3.1"
pymongo = "^4.3.3"
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]